+++ bindings/python/src/rustuary/__init__.py
@@ -5,6 +5,7 @@ from .errors import ColumnMappingError
 from .mapping import ClaimsMapping, ExposureMapping
 from .metadata import ModelRunMetadata
 from .triangle import Triangle
+from .triangle_definition import SegmentDefinition, TriangleDefinition
 
 __all__ = [
     "ChainLadder",
@@ -13,5 +14,7 @@ __all__ = [
     "ExposureMapping",
     "ModelRunMetadata",
     "ReserveResult",
+    "SegmentDefinition",
     "Triangle",
+    "TriangleDefinition",
 ]